chore: refactor jobsd group functions
